+++ R/thin.bvecmodel.R
@@ -34,32 +34,12 @@
 #' @export
 #' @method thin bvecmodel
 thin.bvecmodel <- function(x, thin = 10, ...) {
-  
+
   draws <- nrow(x[["posterior"]][["u_sigma_inv"]][["coeffs"]])
   pos_thin <- seq(from = thin, to = draws, by = thin)
-  start <- pos_thin[1]
-  end <- pos_thin[length(pos_thin)]
-  
-  # Posteriors with sub-lists
-  vars <- c("a", "beta", "psi", "u_sigma_inv", "u_omega_inv")
-  subvars <- c("coeffs", "lambda")
-  for (i in vars) {
-    if (!is.null(x[["posterior"]][[i]])){
-      for (j in subvars) {
-        if (!is.null(x[["posterior"]][[i]][[j]])) {
-          x[["posterior"]][[i]][[j]] <- coda::mcmc(as.matrix(x[["posterior"]][[i]][[j]][pos_thin,]), start = start, end = end, thin = thin)
-        }
-      } 
-    }
-  }
-  
-  # Posteriors w/o sub-lists
-  vars <- c("loglik", "forecast", "forecast_errors")
-  for (i in vars) {
-    if (!is.null(x[["posterior"]][[i]])){
-      x[["posterior"]][[i]] <- coda::mcmc(as.matrix(x[["posterior"]][[i]][pos_thin,]), start = start, end = end, thin = thin)
-    }
-  }
+  # The helper is the VAR method's; see .thin_draws() for why it names no
+  # elements.
+  x[["posterior"]] <- .thin_draws(x[["posterior"]], pos_thin, draws, thin)
 
   return(x)
 }